+++ pyprom/__init__.py
@@ -1,6 +1,6 @@
 """PyProm: This library includes tools for surface network analysis.
 """
-version_info = (0, 1, 1)
+version_info = (0, 2, 0)
 __name__ = 'pyProm'
 __doc__ = 'A python surface network analysis script'
 __author__ = 'Marc Howes'